Started AVLTree Implementation
Finished implementing all of the required functions. I was experiencing a fucking wild infinite recursion error, but I realized that this was due to the rebalance function receiving nodes with in the balance factor range of (-1, 0, 1) so I decided to balance inside of each rotation manually rather than call the function.