main almost done
Implemented the main function, that handles the simulation of the airport. Fixed the use of the auto iterator in 2 loops because it was confusing me during debugging. Removed the debugging formatted output from the display ac method.
Current Issues:
when testing on t31.in
Expected output:
Entering simulation
		0 1 departing 3
		0 2 arriving 3
		0 3 departing 3
		0 4 arriving 4
		0 5 departing 4
		0 6 arriving 2
		0 7 arriving 3
	Runway A
		0 1 departing 3
	Runway B
		0 6 arriving 2
Actual Output:
Time step 0
	Entering simulation
		0 1 departing 3
		0 2 arriving 3
		0 3 departing 3
		0 4 arriving 4
		0 5 departing 4
		0 6 arriving 2
		0 7 arriving 3
	Runway A
		0 5 departing 4
	Runway B
		0 4 arriving 4
So it looks like the id is having an issue. Going to do some digging I suspect its something to do with the currentAircraft obj.