Fixed Priority Issue
I fixed the issue with the planes displaying improperly compared to the expected output. Just based on checking some of the shorter tests it worked, but I am going to write a python program to run it through all of the tests like it does on ingenious because I need a 100 on this project.