Final Version All Tests Pass
Finally fixed the issues that I was running into with the overloaded operators, and fixed the runway loading logic to properly handle output.
Modified Functions:
bool operator>(): Added more explicit if statement logic, also comparing the priorities if they are not equal rather than equal.
bool operator<(): Added same logic as the > operator just with sign reversed.
Also cleaned up the pq class so that I would not cerr out as it was messing with my testing.
int main(): Changed the logic of how pq runways are handled. Rather than using if/elif for each runway processing just one per runway and now one if statement to catch edge case of runwayB being unused and there are still departures. Essentially just changed the runway priorities.

Added Files/directories:
New directory: testingrootdir - I used this directory to store all of the project tests in and for all testing purposes. Inside is also a python program with the name test_runner.py. This is a simple python program I wrote in order to do more efficient targeted testing. It uses the os, subprocess, and difflib to first compile the c++ program. It then takes the test_cases directory which has all of the provided test cases and runs each input through the compiled airportdriver program. It compares the program output with the .out for each test case. If the test passed it just outputs pass, if the test failed it uses difflib to output where the actual output differs from the expected output. The final output is the amount of tests ran, the amount passed and the amount failed. With all file versions here, this is the output of the test_runner.py program:
Test Summary:
Total Tests Run: 95
Tests Passed: 95
Tests Failed: 0
So based on the testing the airportdriver.cpp program works perfectly :)